Added FieldMap

+++ GameServer/ServerCoreLibrary/ServerCoreLibrary/PacketHeader.h
@@ -1,8 +1,6 @@
 #pragma once
-struct PacketHeader
-{
-	uint16 size;
-	uint16 prefix;
-	uint16 id; // 프로토콜ID (ex. 1=로그인, 2=이동요청)
+struct PacketHeader {
+  int size;
+  int prefix;
+  int id;  // 프로토콜ID (ex. 1=로그인, 2=이동요청)
 };
-
